bot/requirements.txt
aiogram==3.20.0.post0
asyncpg
python-dotenv